Implement option filtering for input dropdown
